Improve handling of host tools
Summary:
We now no longer need the corresponding native variants of the frameworks
added as explicit dependencies, the target ones take care of this. The
approach is inspired by how the Wayland recipe handles its code generator.

This also uncovered an issue with KAuth, which can't build its host tools
yet. This uses the target version as a placeholder for now, to not block
the build of everything else.

Reviewers: cordlandwehr

Reviewed By: cordlandwehr

Differential Revision: https://phabricator.kde.org/D7023