Fix variable expansion in conditional overrides
When using conditional override syntax, this should only be combined
with the append/prepend/remove operators. This commit rewrites the
claas-native and class-target overrides such that it does not replace
any recipe set values of EXTRA_OECMAKE (reading the removed syntax it is
very counter-intuitive, but sadly really happens). The problem was:

EXTRA_OECMAKE:class-target += " foo" # from class
EXTRA_OECMAKE += " baa" # from recipe

to be evaluated to EXTRA_OECMAKE = " foo". By this change, the values
are appended as expected, even if the recipe is not using conditional
overrides for the value assignment.